requirements.txt
yt-dlp>=2023.8.3
openai-whisper>=20221106
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
ffmpeg-python>=0.2.0
faiss-cpu>=1.7.3
sentence-transformers>=2.2.2
numpy>=1.23.5
transformers>=4.28.1
openai>=0.27.8
streamlit>=1.24.1
nltk>=3.8
